requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
pylint
python-dotenv>=0.5.1
